requirements.txt
fastapi
google-api-core==2.15.0
google-api-python-client==2.111.0
google-auth==2.25.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.62.0
Jinja2==3.1.2
uvicorn==0.24.0.post1
